requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.11.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.12.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==3.28.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cyclopts==3.22.5
dash==3.1.1
debugpy==1.8.14
decorator==5.2.1
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastmcp==2.11.2
Flask==3.1.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.6
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Js2Py==0.74
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
kombu==5.6.2
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.7.0
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.3.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==24.0
pandas==2.3.0
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pillow==12.0.0
pipwin==0.5.2
platformdirs==4.3.7
plotly==6.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.5
pyperclip==1.9.0
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
redis==7.1.0
referencing==0.36.2
reportlab==4.4.5
requests==2.31.0
retrying==1.4.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.3
sse-starlette==3.0.2
sseclient==0.0.27
stack-data==0.6.3
starlette==0.47.2
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.3
tzlocal==5.2
urllib3==2.2.1
uv==0.9.15
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.1
zipp==3.23.0
